src/railway_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
itsdangerous==2.1.2
PyJWT==2.8.0
pymongo==4.6.0
redis==5.0.1
websockets==12.0
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
psutil==5.9.6 